fix: preserve Nuxt internal error route
